requirements.txt
Augmentor==0.2.8
autograd==1.3
cloudpickle==1.3.0
cupy-cuda101
fastprogress==0.2.3
graphviz==0.10.1
image==1.5.32
imageio==2.4.1
imagesize==1.2.0
librosa==0.6.3
lmdb==0.98
matplotlib==3.2.2
matplotlib-venn==0.11.5
np-utils==0.5.12.1
numba==0.48.0
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
openpyxl==2.5.9
pathlib==1.0.1
Pillow==8.2.0
pip-tools==4.5.1
pydot==1.3.0
pydot-ng==2.0.0
pydotplus==2.0.2
pygame
pyglet
Pygments
regex
requests
requests-oauthlib
scikit-image
scikit-learn
scipy==1.4.1
setuptools-git
torch
torchsummary
torchtext
torchvision
tornado
tqdm
urllib3
